requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
waitress==2.1.2
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2